refactor: tidy up types for `/parties` route
