--- src/lib/partyRequirements/interests.ts
@@ -1,31 +1,0 @@
-export const interests = [
-	'Beginner Friendly',
-	'Web',
-	'Databases',
-	'Machine Learning',
-	'Cybersecurity',
-	'Mobile Development',
-	'Game Development',
-	'Cooking',
-	'UI/UX',
-	'Fitness',
-	'Mobile',
-	'Health',
-	'Art',
-	'Design',
-	'Languages',
-	'Education',
-	'Community',
-	'Sustainability',
-	'App Development',
-	'Artisan Crafts',
-	'Web Development',
-	'E-commerce',
-	'Mental Health',
-	'AI',
-	'Chatbot Development',
-	'Environment',
-	'Virtual Reality',
-	'History'
-] as const;
-export type Interest = (typeof interests)[number];
