+++ src/lib/types/PartyRequirement.ts
@@ -1,0 +1,5 @@
+export type PartyRequirement = {
+	interests: string[];
+	commitment: string;
+	skills: string[];
+};
