+++ src/lib/types/index.ts
@@ -1,2 +1,3 @@
 export * from './Person';
 export * from './MessageFeed';
+export * from './Party';
